+++ app/src/androidTest/java/com/mocomoco/tradin/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.tradin
+package com.mocomoco.tradin
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
