--- app/src/main/java/TradInApplication.kt
@@ -1,6 +1,0 @@
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class TradInApplication {
-
-}
