--- app/src/main/java/com/example/tradin/ui/TradInApp.kt
@@ -1,15 +1,0 @@
-package com.example.tradin.ui
-
-import androidx.compose.runtime.Composable
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.rememberNavController
-
-@Composable
-fun TradInApp() {
-    val navController = rememberNavController()
-
-    NavHost(navController = navController, startDestination = "") {
-
-    }
-
-}
