--- app/src/main/java/com/example/tradin/ui/home/HomeViewModel.kt
@@ -1,9 +1,0 @@
-package com.example.tradin.ui.home
-
-import base.BaseViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-
-@HiltViewModel
-class HomeViewModel : BaseViewModel() {
-
-}
