+++ app/src/main/java/com/mocomoco/tradin/TradInApplication.kt
@@ -1,0 +1,7 @@
+package com.mocomoco.tradin
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class TradInApplication : Application()
