+++ app/src/main/java/com/mocomoco/tradin/ui/MainActivity.kt
@@ -1,0 +1,27 @@
+package com.mocomoco.tradin.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.viewModels
+import androidx.core.splashscreen.SplashScreen
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import com.mocomoco.tradin.ui.theme.TradInTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    private val viewModel: MainViewModel by viewModels()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        installSplashScreen().apply {
+            this.setKeepOnScreenCondition(SplashScreen.KeepOnScreenCondition { viewModel.loading.value })
+        }
+        setContent {
+            TradInTheme {
+               MainScreen()
+            }
+        }
+    }
+}
