+++ app/src/main/java/com/mocomoco/tradin/ui/MainScreen.kt
@@ -1,0 +1,14 @@
+package com.mocomoco.tradin.ui
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.compose.rememberNavController
+
+@Composable
+fun MainScreen() {
+    val navController = rememberNavController()
+
+    TradInNavGraph(
+        navController = navController,
+        startDestination = TradInDestinations.HOME_ROUTE
+    )
+}
