+++ app/src/main/java/com/mocomoco/tradin/ui/MainViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.tradin.ui
+package com.mocomoco.tradin.ui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -7,9 +7,10 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
+import javax.inject.Inject
 
 @HiltViewModel
-class MainViewModel: ViewModel() {
+class MainViewModel @Inject constructor(): ViewModel() {
 
     private val _loading = MutableStateFlow(true)
     val loading = _loading.asStateFlow()
