+++ app/src/main/java/com/mocomoco/tradin/ui/home/HomeScreen.kt
@@ -1,5 +1,6 @@
-package com.example.tradin.ui.home
+package com.mocomoco.tradin.ui.home
 
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.viewmodel.compose.viewModel
 
@@ -7,5 +8,5 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 fun HomeScreen(
     viewModel: HomeViewModel = viewModel()
 ) {
-
+    Text(text = viewModel.a)
 }
