+++ app/src/main/java/com/mocomoco/tradin/ui/home/HomeViewModel.kt
@@ -1,0 +1,10 @@
+package com.mocomoco.tradin.ui.home
+
+import base.BaseViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class HomeViewModel @Inject constructor() : BaseViewModel() {
+    val a = "asdf"
+}
