+++ app/src/main/java/com/mocomoco/tradin/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.example.tradin.ui.theme
+package com.mocomoco.tradin.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
